test: user controller tests
